+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Update :meth:`pycov.CovFrame.from_bam` method to use filename as sample name when the SM tag is missing.
 * Add new method :meth:`pyvcf.row_phased`. From now on, it's used to get the ``pyvcf.VcfFrame.phased`` property.
 * Add new method :meth:`pyvcf.split` and :command:`vcf-split` for splitting VCF by individual.
+* Update :meth:`pyvcf.merge` method, :meth:`pyvcf.VcfFrame.merge` method, and :command:`vcf-merge` command to automatically handle the 'chr' string.
 
 0.29.0 (2021-12-19)
 -------------------